+++ README.md
@@ -1,12 +1,12 @@
-# RPGram
+# rpgram.RPGram
 
 ## Dependencies
 - [TelegramBots](https://mvnrepository.com/artifact/org.telegram/telegrambots)
 
 ## Building from sources
 ```
-$ git clone https://github.com/Zemllia/RPGram.git
-$ cd RPGram
+$ git clone https://github.com/Zemllia/rpgram.RPGram.git
+$ cd rpgram.RPGram
 $ mvn compile
 ```
