+++ RPGRAMSERVER.xml
@@ -99,7 +99,7 @@
     <pathelement location="${path.variable.maven_repository}/com/fasterxml/jackson/core/jackson-annotations/2.9.9/jackson-annotations-2.9.9.jar"/>
   </path>
   
-  <path id="library.maven:_com.fasterxml.jackson.core:jackson-core:2.9.9.classpath">
+  <path id="library.maven:_com.fasterxml.jackson.core:jackson-rpgram.core:2.9.9.classpath">
     <pathelement location="${path.variable.maven_repository}/com/fasterxml/jackson/core/jackson-core/2.9.9/jackson-core-2.9.9.jar"/>
   </path>
   
@@ -211,7 +211,7 @@
     <pathelement location="${path.variable.maven_repository}/org/apache/logging/log4j/log4j-api/2.12.0/log4j-api-2.12.0.jar"/>
   </path>
   
-  <path id="library.maven:_org.apache.logging.log4j:log4j-core:2.12.0.classpath">
+  <path id="library.maven:_org.apache.logging.log4j:log4j-rpgram.core:2.12.0.classpath">
     <pathelement location="${path.variable.maven_repository}/org/apache/logging/log4j/log4j-core/2.12.0/log4j-core-2.12.0.jar"/>
   </path>
   
@@ -291,7 +291,7 @@
     <pathelement location="${path.variable.maven_repository}/org/glassfish/jersey/media/jersey-media-json-jackson/2.29/jersey-media-json-jackson-2.29.jar"/>
   </path>
   
-  <path id="library.maven:_org.hamcrest:hamcrest-core:1.1.classpath">
+  <path id="library.maven:_org.hamcrest:hamcrest-rpgram.core:1.1.classpath">
     <pathelement location="${path.variable.maven_repository}/org/hamcrest/hamcrest-core/1.1/hamcrest-core-1.1.jar"/>
   </path>
