+++ src/main/java/rpgram/DialogEngine.java
