+++ src/main/java/rpgram/NPC.java
