+++ src/main/java/rpgram/Treasure.java
