+++ src/main/java/rpgram/core/GameObject.java
