+++ src/main/java/rpgram/core/NamesGenerator.java
