+++ src/main/java/rpgram/core/Position.java
