+++ src/main/java/rpgram/core/utils/Random.java
