+++ src/main/java/rpgram/maps/MapLayers.java
@@ -1,0 +1,10 @@
+package rpgram;
+
+public enum MapLayers {
+    WEIGHTS,
+    CAVES,
+    GROUND,
+    ENVIRONMENT,
+    PLAYERS,
+    HEAVENS
+}
